remove emr instance fleet resource
